requirements.txt
requests
python-telegram-bot
gunicorn
python-dotenv